Implemented
a) an option to specify oscillator frequencies for the nil8 initialisation: practical for deformed initializations.
b) Routines to calculate the single-particle matrix elements of all angular momenta combinations (and their printing).
c) Calculation of <J^2> and Belyaev moment of inertia for HF and HF+BCS calculations.
d) Removed some trailing debris from various places.
e) Calculation of the rigid body moment of inertia.
f) Bugfixed the calculation of the particle number dispersion: a factor of two was missing.
g) Printing of the single-particle spwfs is now more in line with MOCCa.