Made support for higher-level codes possible
a) The main routine of Tantalus is now a subroutine.
b) Said subroutine can be called with optional input arguments, detailing the way Tantalus reads its input.
   These arguments are optional, when absent the input is read from STDIN.
c) The "ordinary" Tantalus program is now simply encapsulated in the run_single file. It is virtually empty
   and simply calls the subroutine.
d) More advanced runs can now be coded by writing your own run_WHATEVER.f90. As an example, there is a
  run_mpi.f90 which is a basic example.
e) Said changes are reflected in the Makefile which can now be used as
	'make' or 'make single' => Ordinary tantalus, with run_single.f90
	'make mpi'              => MPI-version, with run_mpi.f90