Fixed the incorrect self-consistency of neutron charge form factor effects
Previously, the neutrons did not feel the Coulomb potential. This is fine
when neutrons are point particles, but when they have a charge form factor
they should feel the Coulomb potential (folded with an appropriate
form factor). This feedback was not coded when the user tried to include
the neutron charge form factor selfconsistently.

Test: Pb208, HF, GSk1 (without rotational correction)
       Total energy:   -1636.283261
         from spwfs:   -1636.283262

The old option is for now still included, through the new input parameter
	neutroncoulomberror = .true./.false.
which can be used in forces.param. If set to .true. the code also prints
a deliberately ugly warning when you do that.