Added several extra "advanced output" options, triggered by relevant keywords in the /IO/ namelist. Each of these tell the code to write a newfile (with name determined by the keyword) at the end of the iterations/
  *) BXLFIT :   like before, triggers a one-line file for use with
                the fitting codes in the Brussels group
  *) denfile:   a file containing full 3D information on neutron,
                proton and charge density
  *) potfile:   a file containing full information on the central
                potential (neutron and proton), coulomb potential,
	        spin-orbit potential (neutron and proton) and
                pair potential (neutron and proton). Both so and
	        pair potentials are currently written as zero.
  *) SPHFFILE:  a file containing information on the single-particle
	        states in the Hartree-Fock basis.
  *) SPCAnFILE: a file containing information on the single-particle
	        states in the basis that diagonalizes RHO (which
	        might or might not be the canonical basis).

All of these files should be readable directly with Numpy, as well
as many other plotting programs. They contain headers that should
allow for the labelling/reconstruction of calculations, all
starting with '#'-symbols, such that automated processing of this
information will take a little scripting.

For detailed information on the content of each file, see the
headers of these files and the routines that create them in
the source file IO.f90.

This implementation was accompanied by a small restructuring of
the output routines, as well as a bugfix of the printing of the
runtime parameter checkpointiter.