Bugfix: identification of canonical partners
Before, the identification of the canonical partners in the canonical basis
was somewhat ad-hoc by comparing occupation factors with a large precision.
In pseudocode:

   if(|rho(i) - rho(j)| < 1e-12 and kappa(i,j) > 1e-14 ) then
	partner(i) = j
        partner(j) = i
   endif

For an as yet unknown reason, this failed for cranking calculations, as the
occupation factors are almost never that close, even for small values of
omega.

Instead, the code now relies on a 'minimize difference' method. In pseudocode:

   partner(i) = minarg [ abs(rho(i) - rho(j) | j = 1, N, eta_j != eta_i]

i.e. the code looks for the level(i) with the canonical occupation that is
closest to the original one (i).

This strategy is different from the one used in Esperance, which I don't like
that much as it depends on cutoff parameters (like the 1e-12 and 1e-14 above).
This new method is parameter free.