Bugfix: out-of-bounds error in the identification of partner qps
In the find_partner subroutine in HFB_direct.f90, the identification
of the qp that is the closest time-reversal partner was wrong.

Before:
      tr_qp(       1:  N     ) = + Bogo(1:  N+N2+1,column)
      tr_qp(  N   +1:  N+  N2) = - Bogo(1:2*N+N2+1,column)
      tr_qp(  N+N2+1:2*N+  N2) =   0
      tr_qp(2*N+N2+1:2*N+2*N2) =   0

where on careful inspection one realizes that the arrays in the
first two lines have different sizes.

After:
      tr_qp(       1:  N     ) = + Bogo(  N+N2+1:2*N+  N2,column)
      tr_qp(  N   +1:  N+  N2) = - Bogo(2*N+N2+1:2*N+2*N2,column)
      tr_qp(  N+N2+1:2*N+  N2) =   0
      tr_qp(2*N+N2+1:2*N+2*N2) =   0
                               ! Sign due to time-reversal

This showed up as segfaults on some systems, some of the time
(but never, seemingly, on mine...)