Bugfix: statements generated with too much "Der_" prefixes
On input of terms that combined laplacian and derivatives, such as
  E_Derk_D_I_Sm_Derk_Lap_D_I_Sm
the code generated isospin recoupling statements with too much
derivatives, such as
  der_Der_Der_D_I_S(:,.....,3) = der_Der_Der_D_I_S(:,.....,1) + der_Der_Der_D_I_S(:,.....,2)
while no der_Der_Der_D_I_S object was defined.

The issue is that the code accidentally used the "iso_der" template line when
trying to generate code calculating actually
   Lap_Der_D_I_S(:,......,3)  = Lap_Der_D_I_S(:,......,1) + Lap_Der_D_I_S(:,......,2)
i.e. the logic was right, but the code filled in the indices in the wrong string template.
This was a three-character fix....

I also rendered the capitalization more consistent.