Bugfix in ParseOperators
The ParseOperators routine was not able to distinguish

   C_NkNk_NS vs C_NN_NoSo

in the functional. Both got parsed as

  left  = 'NN'
  right = 'NS'
  coupling = [(0,1)]

which is wrong for the second one.

The origin was the lack of accounting for operators that were not coupled
to other ones. This was not foreseen in the very earliest versions of the
code and only later added as use-case.