Inclusion of the Ylm decomposition program
auxiliary/decomposition/ now contains a fully functional and tested
program for the calculation of the expansion coefficients of densities
in terms of spherical harmonics. It is written in FORTRAN for speed
and thus needs to be compiled separately, since the integrations can
be somewhat demanding.

Please see the manual "decomp.pdf" for a discussion of conventions
of these coefficients and a small manual of operation.

This FORTRAN code is able to completely replace the PYTHON scripting
I had before. This will be deleted in a later commit, just to keep
some track of things and possibly eliminate bugs that remain to be
detected.