Graceful exit for the Fermi solver in the HFB_gradient module
If the particle number was found to be a non-monotone function of the
Fermi energy, the gradient solver really could not handle that in the
previous version of the code.

The new version handles this much more gracefully: if such a problem is
detected, the solver now no longer tries to ENFORCE the particle
number during the gradient step. Instead, it ignores such errors and
proceeds to step anyway but then adds a corrective step AFTER the
initial gradient step. This seems to be much better behaved in initial
tests.

However, if problems would persist, the code now fails gracefully and
announces a problem with the convergence message "FERMI" for the
BXL output.

Also, the estimation of the gradient_stepsize and gradient_mu are now
more careful in the very first iteration.